requirements.txt
django==2.2.6
pytz==2020.4              # via django
sqlparse==0.4.1           # via django

sorl-thumbnail==12.6.3    # images
pillow==7.0.0             # images

django-environ            #.env file manager

psycopg2-binary           # database engine